+++ packages/common/src/PublicNetworks.ts
@@ -78,7 +78,7 @@ export const PublicNetworks: PublicNetworksType = {
   },
   42161: { name: "arbitrum", etherscan: "https://arbiscan.io" },
   43114: { name: "avalanche", etherscan: "https://snowtrace.io" },
-  8453: { name: "base", etherscan: "https://goerli.basescan.org" },
+  8453: { name: "base", etherscan: "https://basescan.org" },
   84531: { name: "base-goerli", etherscan: "https://goerli.basescan.org" },
   421611: { name: "arbitrum-rinkeby", etherscan: "https://testnet.arbiscan.io" },
   421613: { name: "arbitrum-goerli", etherscan: "https://goerli.arbiscan.io" },