+++ packages/financial-templates-lib/src/logger/PagerDutyV2Transport.ts
@@ -3,6 +3,8 @@ import Transport from "winston-transport";
 import { event } from "@pagerduty/pdjs";
 import * as ss from "superstruct";
 
+import { TransportError } from "./TransportError";
+
 type TransportOptions = ConstructorParameters<typeof Transport>[0];
 export type Severity = "critical" | "error" | "warning" | "info";
 export type Action = "trigger" | "acknowledge" | "resolve";
@@ -10,11 +12,13 @@ export type Action = "trigger" | "acknowledge" | "resolve";
 const Config = ss.object({
   integrationKey: ss.string(),
   customServices: ss.optional(ss.record(ss.string(), ss.string())),
+  logTransportErrors: ss.optional(ss.boolean()),
 });
 // Config object becomes a type
 // {
 //   integrationKey: string;
 //   customServices?: Record<string,string>;
+//   logTransportErrors?: boolean;
 // }
 export type Config = ss.Infer<typeof Config>;
 
@@ -26,10 +30,15 @@ export function createConfig(config: unknown): Config {
 export class PagerDutyV2Transport extends Transport {
   private readonly integrationKey: string;
   private readonly customServices: { [key: string]: string };
-  constructor(winstonOpts: TransportOptions, { integrationKey, customServices = {} }: Config) {
+  public readonly logTransportErrors: boolean;
+  constructor(
+    winstonOpts: TransportOptions,
+    { integrationKey, customServices = {}, logTransportErrors = false }: Config
+  ) {
     super(winstonOpts);
     this.integrationKey = integrationKey;
     this.customServices = customServices;
+    this.logTransportErrors = logTransportErrors;
   }
   // pd v2 severity only supports critical, error, warning or info.
   public static convertLevelToSeverity(level?: string): Severity {
@@ -39,7 +48,7 @@ export class PagerDutyV2Transport extends Transport {
     return "error";
   }
   // Note: info must be any because that's what the base class uses.
-  async log(info: any, callback: () => void): Promise<void> {
+  async log(info: any, callback: (error?: unknown) => void): Promise<void> {
     try {
       // we route to different pd services using the integration key (routing_key), or multiple services with the custom services object
       const routing_key = this.customServices[info.notificationPath] ?? this.integrationKey;
@@ -57,6 +66,8 @@ export class PagerDutyV2Transport extends Transport {
         },
       });
     } catch (error) {
+      // We don't want to emit error if this same transport is used to log transport errors to avoid recursion.
+      if (!this.logTransportErrors) return callback(new TransportError("PagerDuty V2", error, info));
       console.error("PagerDuty v2 error", error);
     }