+++ packages/monitor-v2/src/monitor-og/oSnapAutomation.ts
@@ -3,8 +3,10 @@ import type { Signer } from "@ethersproject/abstract-signer";
 import { ERC20Ethers } from "@uma/contracts-node";
 import {
   ProposalDeletedEvent,
+  ProposalExecutedEvent,
   TransactionsProposedEvent,
 } from "@uma/contracts-node/typechain/core/ethers/OptimisticGovernor";
+import { createEtherscanLinkMarkdown } from "@uma/common";
 import assert from "assert";
 import retry, { Options as RetryOptions } from "async-retry";
 import { ContractReceipt, utils as ethersUtils } from "ethers";
@@ -13,12 +15,13 @@ import { gql } from "graphql-tag";
 
 import { getEventTopic } from "../utils/contracts";
 import { createSnapshotProposalLink } from "../utils/logger";
-import { logSubmittedProposal } from "./MonitorLogger";
+import { logSubmittedDispute, logSubmittedExecution, logSubmittedProposal } from "./MonitorLogger";
 
 import {
   getBlockTimestamp,
   getContractInstanceWithProvider,
   getOgByAddress,
+  getOo,
   Logger,
   MonitoringParams,
   runQueryFilter,
@@ -35,6 +38,7 @@ import {
   SafeSnapSafe,
   SnapshotProposalGraphql,
   verifyIpfs,
+  verifyProposal,
   verifyRules,
   verifyVoteOutcome,
 } from "./SnapshotVerification";
@@ -47,6 +51,15 @@ interface SupportedParameters {
 
 interface SupportedModules {
   [ogAddress: string]: SupportedParameters;
+}
+
+export interface SupportedProposal {
+  event: TransactionsProposedEvent;
+  parameters: SupportedParameters;
+}
+
+export interface DisputableProposal extends SupportedProposal {
+  verificationResult: { verified: false; error: string };
 }
 
 // Expanded interface for easier processing of Snapshot proposals. Original Snapshot proposal can contain multiple safes
@@ -159,16 +172,16 @@ const getSupportedSnapshotProposals = async (
   return expandedProposals.filter((proposal) => isSafeSupported(proposal.safe, supportedModules, params.chainId));
 };
 
-// Get all proposals on supported oSnap modules that have not been discarded. Discards are most likely due to disputes,
+// Get all proposals on provided oSnap modules that have not been discarded. Discards are most likely due to disputes,
 // but can also occur on OOv3 upgrades.
 const getUndiscardedProposals = async (
-  supportedModules: SupportedModules,
+  ogAddresses: string[],
   params: MonitoringParams
 ): Promise<Array<TransactionsProposedEvent>> => {
-  // Get all proposals for all supported modules.
+  // Get all proposals for all provided modules.
   const allProposals = (
     await Promise.all(
-      Object.keys(supportedModules).map(async (ogAddress) => {
+      ogAddresses.map(async (ogAddress) => {
         const og = await getOgByAddress(params, ogAddress);
         return runQueryFilter<TransactionsProposedEvent>(og, og.filters.TransactionsProposed(), {
           start: 0,
@@ -178,10 +191,10 @@ const getUndiscardedProposals = async (
     )
   ).flat();
 
-  // Get all deleted proposals for all supported modules.
+  // Get all deleted proposals for all provided modules.
   const deletedProposals = (
     await Promise.all(
-      Object.keys(supportedModules).map(async (ogAddress) => {
+      ogAddresses.map(async (ogAddress) => {
         const og = await getOgByAddress(params, ogAddress);
         return runQueryFilter<ProposalDeletedEvent>(og, og.filters.ProposalDeleted(), {
           start: 0,
@@ -193,8 +206,8 @@ const getUndiscardedProposals = async (
 
   // Filter out all proposals that have been deleted by matching assertionId. assertionId should be sufficient property
   // for filtering as it is derived from module address, transaction content and assertion time among other factors.
-  const deletedAssertionIds = deletedProposals.map((deletedProposal) => deletedProposal.args.assertionId);
-  return allProposals.filter((proposal) => !deletedAssertionIds.includes(proposal.args.assertionId));
+  const deletedAssertionIds = new Set(deletedProposals.map((deletedProposal) => deletedProposal.args.assertionId));
+  return allProposals.filter((proposal) => !deletedAssertionIds.has(proposal.args.assertionId));
 };
 
 // Checks if a safeSnap safe from Snapshot proposal is supported by oSnap automation.
@@ -226,6 +239,29 @@ const filterPotentialProposals = (
     });
     // Exclude Snapshot proposals with matching on-chain proposals
     return matchingOnChainProposals.length === 0;
+  });
+};
+
+// Filters out all Snapshot proposals that cannot be proposed due to blocking on-chain proposals. This is done by
+// matching safe and proposed transactions.
+const filterUnblockedProposals = async (
+  potentialProposals: SnapshotProposalExpanded[],
+  onChainProposals: TransactionsProposedEvent[],
+  params: MonitoringParams
+): Promise<SnapshotProposalExpanded[]> => {
+  // Filter out all on-chain proposals that have been executed since they cannot block new proposals.
+  const unexecutedProposals = await filterUnexecutedProposals(onChainProposals, params);
+
+  return potentialProposals.filter((potentialProposal) => {
+    // Unexecuted proposals with the same safe and matching transactions would block the new proposal.
+    const blockingOnChainProposals = unexecutedProposals.filter((unexecutedProposal) => {
+      return (
+        isMatchingSafe(potentialProposal.safe, params.chainId, unexecutedProposal.address) &&
+        onChainTxsMatchSnapshot(unexecutedProposal, potentialProposal.safe)
+      );
+    });
+    // Exclude Snapshot proposals with blocking on-chain proposals
+    return blockingOnChainProposals.length === 0;
   });
 };
 
@@ -274,6 +310,77 @@ const filterVerifiedProposals = async (
   });
 };
 
+// Filters out all proposals that have been executed on-chain. This results in proposals both before and after their
+// challenge period.
+const filterUnexecutedProposals = async (
+  proposals: TransactionsProposedEvent[],
+  params: MonitoringParams
+): Promise<TransactionsProposedEvent[]> => {
+  // Get all assertion Ids from executed proposals covering modules in input proposals.
+  const executedAssertionIds = new Set(
+    (
+      await Promise.all(
+        Array.from(new Set(proposals.map((proposal) => proposal.address))).map(async (ogAddress) => {
+          const og = await getOgByAddress(params, ogAddress);
+          const executedProposals = await runQueryFilter<ProposalExecutedEvent>(og, og.filters.ProposalExecuted(), {
+            start: 0,
+            end: params.blockRange.end,
+          });
+          return executedProposals.map((executedProposal) => executedProposal.args.assertionId);
+        })
+      )
+    ).flat()
+  );
+
+  // Filter out all proposals that have been executed based on matching assertionId.
+  return proposals.filter((proposal) => !executedAssertionIds.has(proposal.args.assertionId));
+};
+
+// Filter function to check if challenge period has passed for a proposal.
+const hasChallengePeriodEnded = (proposal: TransactionsProposedEvent, timestamp: number): boolean => {
+  return timestamp >= proposal.args.challengeWindowEnds.toNumber();
+};
+
+// Filters supported proposal events and adds their parameters to the result.
+const getSupportedProposals = async (
+  proposals: TransactionsProposedEvent[],
+  params: MonitoringParams
+): Promise<SupportedProposal[]> => {
+  // Get OOv3 for checking if assertion's bond is supported.
+  const oo = await getOo(params);
+
+  // Keep only proposals whose rules are parsable and bond is supported based on its assertionId.
+  const supportedProposals = (
+    await Promise.all(
+      proposals.map(async (event) => {
+        const parsedRules = parseRules(event.args.rules);
+        const { currency, bond } = await oo.getAssertion(event.args.assertionId);
+        const isSupported = parsedRules !== null && isBondSupported(currency, bond.toString(), params.supportedBonds);
+        return isSupported ? { event, parameters: { parsedRules, currency, bond: bond.toString() } } : null;
+      })
+    )
+  ).filter((proposal) => proposal !== null) as SupportedProposal[];
+
+  return supportedProposals;
+};
+
+// Filter proposals that did not pass verification and also retain verification result for logging.
+const getDisputableProposals = async (
+  proposals: SupportedProposal[],
+  params: MonitoringParams
+): Promise<DisputableProposal[]> => {
+  // TODO: We should separately handle IPFS and Graphql server errors. We don't want to submit disputes immediately just
+  // because IPFS gateway or Snapshot backend is down.
+  return (
+    await Promise.all(
+      proposals.map(async (proposal) => {
+        const verificationResult = await verifyProposal(proposal.event, params);
+        return !verificationResult.verified ? { ...proposal, verificationResult } : null;
+      })
+    )
+  ).filter((proposal) => proposal !== null) as DisputableProposal[];
+};
+
 const approveBond = async (
   provider: Provider,
   signer: Signer,
@@ -281,9 +388,6 @@ const approveBond = async (
   bond: string,
   spender: string
 ): Promise<void> => {
-  // If bond is 0, no need to approve.
-  if (bond === "0") return;
-
   // If existing approval matches the bond, no need to proceed.
   const currencyContract = await getContractInstanceWithProvider<ERC20Ethers>("ERC20", provider, currency);
   const currentAllowance = await currencyContract.allowance(await signer.getAddress(), spender);
@@ -292,6 +396,7 @@ const approveBond = async (
   try {
     await (await currencyContract.connect(signer).approve(spender, bond)).wait();
   } catch (error) {
+    // There is no point in proceeding with proposal/dispute if bond approval failed, so we throw an error.
     assert(error instanceof Error, "Unexpected Error type!");
     throw new Error(`Bond approval for ${spender} failed: ${error.message}`);
   }
@@ -323,21 +428,32 @@ const submitProposals = async (
     });
     const explanation = ethersUtils.toUtf8Bytes(proposal.ipfs);
 
+    // Create potential log for simulating/proposing.
+    const proposalAttemptLog = {
+      at: "oSnapAutomation",
+      mrkdwn:
+        "Trying to submit proposal for " +
+        createSnapshotProposalLink(params.snapshotEndpoint, proposal.space.id, proposal.id) +
+        " on oSnap module " +
+        createEtherscanLinkMarkdown(proposal.safe.umaAddress, params.chainId) +
+        " at Snapshot space " +
+        proposal.space.id,
+      notificationPath: "optimistic-governor",
+    };
+
     // Check that proposal submission would succeed.
     try {
       await og.callStatic.proposeTransactions(transactions, explanation, { from: await params.signer.getAddress() });
     } catch (error) {
-      assert(error instanceof Error, "Unexpected Error type!");
-      // TODO: We should separately handle the duplicate proposals error. This can occur if there are multiple proposals
-      // with the same transactions (e.g. funding in same amount tranches split across multiple proposals). We should
-      // only warn when first encountering the blocking proposal and then ignore it in any future runs.
-      throw new Error(
-        `Proposal submission for ${createSnapshotProposalLink(
-          params.snapshotEndpoint,
-          proposal.space.id,
-          proposal.id
-        )} would fail: ${error.message}`
-      );
+      // Log error and proceed with the next proposal.
+      logger.error({ ...proposalAttemptLog, message: "Proposal submission would fail!", error });
+      continue;
+    }
+
+    // If submitting transactions is disabled, log the proposal attempt and proceed with the next proposal.
+    if (!params.submitAutomation) {
+      logger.info({ ...proposalAttemptLog, message: "Proposal transaction would succeed" });
+      continue;
     }
 
     // Submit proposal and get receipt.
@@ -346,14 +462,9 @@ const submitProposals = async (
       const tx = await og.connect(params.signer).proposeTransactions(transactions, explanation);
       receipt = await tx.wait();
     } catch (error) {
-      assert(error instanceof Error, "Unexpected Error type!");
-      throw new Error(
-        `Proposal submission for ${createSnapshotProposalLink(
-          params.snapshotEndpoint,
-          proposal.space.id,
-          proposal.id
-        )} failed: ${error.message}`
-      );
+      // Log error and proceed with the next proposal.
+      logger.error({ ...proposalAttemptLog, message: "Proposal submission failed!", error });
+      continue;
     }
 
     // Log submitted proposal.
@@ -374,6 +485,127 @@ const submitProposals = async (
   }
 };
 
+const submitDisputes = async (logger: typeof Logger, proposals: DisputableProposal[], params: MonitoringParams) => {
+  assert(params.signer !== undefined, "Signer must be set to dispute proposals.");
+  const disputerAddress = await params.signer.getAddress();
+
+  for (const proposal of proposals) {
+    const oo = await getOo(params);
+
+    // Approve bond based on passed proposal parameters.
+    await approveBond(
+      params.provider,
+      params.signer,
+      proposal.parameters.currency,
+      proposal.parameters.bond,
+      oo.address
+    );
+
+    // Create potential log for simulating/disputing.
+    const disputeAttemptLog = {
+      at: "oSnapAutomation",
+      mrkdwn:
+        "Trying to submit dispute on assertionId " +
+        proposal.event.args.assertionId +
+        " related to proposalHash " +
+        proposal.event.args.proposalHash +
+        " posted on oSnap module " +
+        createEtherscanLinkMarkdown(proposal.event.address, params.chainId) +
+        " at Snapshot space " +
+        proposal.parameters.parsedRules.space,
+      notificationPath: "optimistic-governor",
+    };
+
+    // Check that dispute submission would succeed.
+    try {
+      await oo.callStatic.disputeAssertion(proposal.event.args.assertionId, disputerAddress, {
+        from: disputerAddress,
+      });
+    } catch (error) {
+      // Log error and proceed with the next dispute.
+      logger.error({ ...disputeAttemptLog, message: "Dispute submission would fail!", error });
+      continue;
+    }
+
+    // If submitting transactions is disabled, log the dispute attempt and proceed with the next dispute.
+    if (!params.submitAutomation) {
+      logger.info({ ...disputeAttemptLog, message: "Dispute transaction would succeed" });
+      continue;
+    }
+
+    // Submit dispute and get receipt.
+    let receipt: ContractReceipt;
+    try {
+      const tx = await oo.connect(params.signer).disputeAssertion(proposal.event.args.assertionId, disputerAddress);
+      receipt = await tx.wait();
+    } catch (error) {
+      // Log error and proceed with the next dispute.
+      logger.error({ ...disputeAttemptLog, message: "Dispute submission failed!", error });
+      continue;
+    }
+
+    // Log submitted dispute.
+    const disputeEvent = receipt.events?.find((e) => e.event === "AssertionDisputed");
+    assert(disputeEvent !== undefined, "AssertionDisputed event not found.");
+    await logSubmittedDispute(logger, proposal, disputeEvent.transactionHash, params);
+  }
+};
+
+const submitExecutions = async (logger: typeof Logger, proposals: SupportedProposal[], params: MonitoringParams) => {
+  assert(params.signer !== undefined, "Signer must be set to execute proposals.");
+  const executorAddress = await params.signer.getAddress();
+
+  for (const proposal of proposals) {
+    const og = await getOgByAddress(params, proposal.event.address);
+
+    // Create potential log for simulating/executing.
+    const executionAttemptLog = {
+      at: "oSnapAutomation",
+      mrkdwn:
+        "Trying to execute proposal with proposalHash " +
+        proposal.event.args.proposalHash +
+        " posted on oSnap module " +
+        createEtherscanLinkMarkdown(proposal.event.address, params.chainId) +
+        " at Snapshot space " +
+        proposal.parameters.parsedRules.space,
+      notificationPath: "optimistic-governor",
+    };
+
+    // Check that execution submission would succeed.
+    try {
+      await og.callStatic.executeProposal(proposal.event.args.proposal.transactions, { from: executorAddress });
+    } catch (error) {
+      // The execution might revert for various reasons (e.g. insufficient funds in safe, transaction guard blocking or
+      // the module has been unplugged). In most of these cases there is nothing the on-call can do, thus log this at
+      // warn level and proceed with the next execution.
+      logger.info({ ...executionAttemptLog, message: "Proposal execution would fail!", error });
+      continue;
+    }
+
+    // If submitting transactions is disabled, log the execution attempt and proceed with the next execution.
+    if (!params.submitAutomation) {
+      logger.info({ ...executionAttemptLog, message: "Execution transaction would succeed" });
+      continue;
+    }
+
+    // Submit execution and get receipt.
+    let receipt: ContractReceipt;
+    try {
+      const tx = await og.connect(params.signer).executeProposal(proposal.event.args.proposal.transactions);
+      receipt = await tx.wait();
+    } catch (error) {
+      // Log error and proceed with the next execution.
+      logger.error({ ...executionAttemptLog, message: "Proposal execution failed!", error });
+      continue;
+    }
+
+    // Log submitted execution.
+    const executionEvent = receipt.events?.find((e) => e.event === "ProposalExecuted");
+    assert(executionEvent !== undefined, "ProposalExecuted event not found.");
+    await logSubmittedExecution(logger, proposal, executionEvent.transactionHash, params);
+  }
+};
+
 export const proposeTransactions = async (logger: typeof Logger, params: MonitoringParams): Promise<void> => {
   // Get supported modules.
   const supportedModules = await getSupportedModules(params);
@@ -382,12 +614,54 @@ export const proposeTransactions = async (logger: typeof Logger, params: Monitor
   const supportedProposals = await getSupportedSnapshotProposals(supportedModules, params);
 
   // Get all undiscarded on-chain proposals for supported modules.
-  const onChainProposals = await getUndiscardedProposals(supportedModules, params);
+  const onChainProposals = await getUndiscardedProposals(Object.keys(supportedModules), params);
 
   // Filter Snapshot proposals that could potentially be proposed on-chain.
   const potentialProposals = filterPotentialProposals(supportedProposals, onChainProposals, params);
-  const verifiedProposals = await filterVerifiedProposals(potentialProposals, supportedModules, params);
+  const unblockedProposals = await filterUnblockedProposals(potentialProposals, onChainProposals, params);
+  const verifiedProposals = await filterVerifiedProposals(unblockedProposals, supportedModules, params);
 
   // Submit proposals.
   await submitProposals(logger, verifiedProposals, supportedModules, params);
+};
+
+export const disputeProposals = async (logger: typeof Logger, params: MonitoringParams): Promise<void> => {
+  // Get all undiscarded on-chain proposals for all monitored modules.
+  const onChainProposals = await getUndiscardedProposals(params.ogAddresses, params);
+
+  // Filter out all proposals that have been executed on-chain.
+  const unexecutedProposals = await filterUnexecutedProposals(onChainProposals, params);
+
+  // Filter out all proposals that have passed their challenge period.
+  const lastTimestamp = await getBlockTimestamp(params.provider, params.blockRange.end);
+  const liveProposals = unexecutedProposals.filter((proposal) => !hasChallengePeriodEnded(proposal, lastTimestamp));
+
+  // Filter only supported proposals and get their parameters.
+  const supportedProposals = await getSupportedProposals(liveProposals, params);
+
+  // Filter proposals that did not pass verification and also retain verification result for logging.
+  const disputableProposals = await getDisputableProposals(supportedProposals, params);
+
+  // Submit disputes.
+  await submitDisputes(logger, disputableProposals, params);
+};
+
+export const executeProposals = async (logger: typeof Logger, params: MonitoringParams): Promise<void> => {
+  // Get all undiscarded on-chain proposals for all monitored modules.
+  const onChainProposals = await getUndiscardedProposals(params.ogAddresses, params);
+
+  // Filter out all proposals that have been executed on-chain.
+  const unexecutedProposals = await filterUnexecutedProposals(onChainProposals, params);
+
+  // Filter out all proposals that have not passed their challenge period.
+  const lastTimestamp = await getBlockTimestamp(params.provider, params.blockRange.end);
+  const unchallengedProposals = unexecutedProposals.filter((proposal) =>
+    hasChallengePeriodEnded(proposal, lastTimestamp)
+  );
+
+  // Filter only supported proposals and get their parameters.
+  const supportedProposals = await getSupportedProposals(unchallengedProposals, params);
+
+  // Submit executions.
+  await submitExecutions(logger, supportedProposals, params);
 };