+++ packages/monitor-v2/src/price-publisher/index.ts
@@ -1,4 +1,4 @@
-import { delay } from "@uma/financial-templates-lib";
+import { delay, waitForLogger } from "@uma/financial-templates-lib";
 import { BotModes, initMonitoringParams, Logger, startupLogLevel } from "./common";
 import { publishPrices } from "./PublishPrices";
 import { resolvePrices } from "./ResolvePrices";
@@ -32,6 +32,7 @@ async function main() {
       await delay(params.pollingDelay);
     } else {
       await delay(5); // Set a delay to let the transports flush fully.
+      await waitForLogger(logger);
       break;
     }
   }
@@ -49,6 +50,7 @@ main().then(
     });
     // Wait 5 seconds to allow logger to flush.
     await delay(5);
+    await waitForLogger(logger);
     process.exit(1);
   }
 );