+++ packages/monitor-v2/src/price-speed-up/index.ts
@@ -1,4 +1,4 @@
-import { delay } from "@uma/financial-templates-lib";
+import { delay, waitForLogger } from "@uma/financial-templates-lib";
 import { BotModes, initMonitoringParams, Logger, startupLogLevel } from "./common";
 import { speedUpPrices } from "./SpeedUpPriceRequests";
 
@@ -30,6 +30,7 @@ async function main() {
       await delay(params.pollingDelay);
     } else {
       await delay(5); // Set a delay to let the transports flush fully.
+      await waitForLogger(logger);
       break;
     }
   }
@@ -47,6 +48,7 @@ main().then(
     });
     // Wait 5 seconds to allow logger to flush.
     await delay(5);
+    await waitForLogger(logger);
     process.exit(1);
   }
 );