+++ packages/contracts-node/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.17](https://github.com/UMAprotocol/protocol/compare/@uma/contracts-node@0.4.16...@uma/contracts-node@0.4.17) (2023-07-17)
+
+**Note:** Version bump only for package @uma/contracts-node
+
 ## [0.4.16](https://github.com/UMAprotocol/protocol/compare/@uma/contracts-node@0.4.15...@uma/contracts-node@0.4.16) (2023-04-18)
 
 **Note:** Version bump only for package @uma/contracts-node